thrive update: session manager

+++ thriveapp/src/main/java/thrivecodestudio/thrive/thriveapp/SessionStorage/Impl/AbstractInMemoryCacheAdapter.java
@@ -3,17 +3,17 @@ package thrivecodestudio.thrive.thriveapp.SessionStorage.Impl;
 import java.util.concurrent.ConcurrentHashMap;
 
 public abstract class AbstractInMemoryCacheAdapter {
-    protected ConcurrentHashMap<Long, SessionData> inMemoryKVstore; 
+    protected ConcurrentHashMap<Byte, SessionData> inMemoryKVstore; 
 
     public AbstractInMemoryCacheAdapter() {
         this.inMemoryKVstore = new ConcurrentHashMap<>();
     }
 
-    public abstract void storeData(final Long key, final SessionData value);
+    public abstract void storeData(final Byte key, final SessionData value);
 
-    public abstract void deleteKVPairs(final Long key);
+    public abstract void deleteKVPairs(final Byte key);
 
-    public abstract SessionData searchKVPairs(final Long key);
+    public abstract SessionData searchKVPairs(final Byte key);
 
     // TODO => add specific fetching operations
 }
